ASSET-26 Move camelcase model
