+++ libraries/models/asset.py
@@ -1,7 +1,6 @@
 from libraries.models.contract_list import ContractList
 from libraries.models.id import Id
 from libraries.models.image_info import ImageInfo
-from libraries.utils.model import CamelCaseModel
 from libraries.models.reference_list import ReferenceList
 from libraries.models.tag_list import TagList
 from libraries.models.templates.camelcase_model import CamelCaseModel
