+++ libraries/models/contract.py
@@ -2,8 +2,8 @@ from pydantic import NonNegativeInt
 
 from libraries.models.address import Address
 from libraries.models.id import Id
-from libraries.utils.model import CamelCaseModel
 from libraries.models.tag_list import TagList
+from libraries.models.templates.camelcase_model import CamelCaseModel
 
 
 class Contract(CamelCaseModel):
